+++ src/index.tsx
@@ -1,4 +1,4 @@
-import { ApolloClient, InMemoryCache, ApolloProvider } from '@apollo/client'
+import { ApolloProvider } from '@apollo/client'
 import { CssBaseline, ThemeProvider } from '@mui/material'
 import { setFilterStore } from '@promoboxx/use-filter/dist/store'
 import localStorageStore from '@promoboxx/use-filter/dist/store/localStorageStore'
@@ -7,14 +7,7 @@ import { createRoot } from 'react-dom/client'
 
 import App from './app/App'
 import theme from './app/theme'
-import tauriGraphqlApolloLink from './tauri/tauriGraphqlApolloLink'
-// import tauriGraphqlHttpLink from './tauri/tauriGraphqlHttpLink'
-
-const client = new ApolloClient({
-  cache: new InMemoryCache(),
-  link: tauriGraphqlApolloLink,
-  // link: tauriGraphqlHttpLink,
-})
+import graphqlClient from './graphql/graphqlClient'
 
 setFilterStore(localStorageStore)
 
@@ -25,7 +18,7 @@ if (!rootElement) {
 }
 
 createRoot(rootElement).render(
-  <ApolloProvider client={client}>
+  <ApolloProvider client={graphqlClient}>
     <StrictMode>
       <ThemeProvider theme={theme}>
         <CssBaseline>
